+++ app/src/main/java/com/techyourchance/dagger2course/common/composition/ActivityCompositionRoot.kt
@@ -1,5 +1,6 @@
 package com.techyourchance.dagger2course.common.composition
 
+import android.view.LayoutInflater
 import androidx.appcompat.app.AppCompatActivity
 import com.techyourchance.dagger2course.questions.FetchQuestionDetailsUseCase
 import com.techyourchance.dagger2course.questions.FetchQuestionsUseCase
@@ -14,6 +15,8 @@ class ActivityCompositionRoot(
     val screensNavigator by lazy {
         ScreensNavigator(activity)
     }
+
+    private val layoutInflater get() = LayoutInflater.from(activity)
 
     private val fragmentManager get() = activity.supportFragmentManager
