+++ app/src/main/java/com/techyourchance/dagger2course/screens/common/fragments/BaseFragment.kt
@@ -1,0 +1,9 @@
+package com.techyourchance.dagger2course.screens.common.fragments
+
+import androidx.fragment.app.Fragment
+import com.techyourchance.dagger2course.screens.common.activities.BaseActivity
+
+open class BaseFragment: Fragment() {
+
+    protected val compositionRoot get() = (requireActivity() as BaseActivity).compositionRoot
+}
