+++ app/src/main/java/com/techyourchance/dagger2course/screens/questionslist/QuestionsListActivity.kt
@@ -1,84 +1,21 @@
 package com.techyourchance.dagger2course.screens.questionslist
 
 import android.os.Bundle
-import android.view.LayoutInflater
-import com.techyourchance.dagger2course.questions.FetchQuestionsUseCase
-import com.techyourchance.dagger2course.questions.Question
-import com.techyourchance.dagger2course.screens.common.ScreensNavigator
+import com.techyourchance.dagger2course.R
 import com.techyourchance.dagger2course.screens.common.activities.BaseActivity
-import com.techyourchance.dagger2course.screens.common.dialogs.DialogsNavigator
-import kotlinx.coroutines.*
 
-class QuestionsListActivity : BaseActivity(), QuestionsListViewMvc.Listener {
-
-    private val coroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Main.immediate)
-
-    private var isDataLoaded = false
-
-    private lateinit var viewMvc: QuestionsListViewMvc
-
-    private lateinit var fetchQuestionsUseCase: FetchQuestionsUseCase
-
-    private lateinit var dialogsNavigator: DialogsNavigator
-
-    private lateinit var screensNavigator: ScreensNavigator
+class QuestionsListActivity : BaseActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
+        setContentView(R.layout.layout_frame)
 
-        viewMvc = QuestionsListViewMvc(LayoutInflater.from(this), null)
-
-        setContentView(viewMvc.rootView)
-
-        fetchQuestionsUseCase = compositionRoot.fetchQuestionsUseCase
-
-        dialogsNavigator = compositionRoot.dialogsNavigator
-
-        screensNavigator = compositionRoot.screensNavigator
-    }
-
-    override fun onStart() {
-        super.onStart()
-        viewMvc.registerListener(this)
-        if (!isDataLoaded) {
-            fetchQuestions()
+        if (savedInstanceState == null) {
+            supportFragmentManager.beginTransaction()
+                    .add(R.id.frame_content, QuestionsListFragment())
+                    .commit()
         }
-    }
 
-    override fun onStop() {
-        super.onStop()
-        coroutineScope.coroutineContext.cancelChildren()
-        viewMvc.unregisterListener(this)
-    }
-
-    override fun onRefreshClicked() {
-        fetchQuestions()
-    }
-
-    private fun fetchQuestions() {
-        coroutineScope.launch {
-            viewMvc.showProgressIndication()
-            try {
-                val result = fetchQuestionsUseCase.fetchLatestQuestions()
-                when (result) {
-                    is FetchQuestionsUseCase.Result.Success -> {
-                        viewMvc.bindQuestions(result.questions)
-                        isDataLoaded = true
-                    }
-                    is FetchQuestionsUseCase.Result.Failure -> onFetchFailed()
-                }
-            } finally {
-                viewMvc.hideProgressIndication()
-            }
-        }
-    }
-
-    private fun onFetchFailed() {
-        dialogsNavigator.showServerErrorDialog()
-    }
-
-    override fun onQuestionClicked(clickedQuestion: Question) {
-        screensNavigator.toQuestionDetails(clickedQuestion.id)
     }
