+++ app/src/main/java/com/techyourchance/dagger2course/screens/questionslist/QuestionsListFragment.kt
@@ -1,0 +1,88 @@
+package com.techyourchance.dagger2course.screens.questionslist
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.techyourchance.dagger2course.questions.FetchQuestionsUseCase
+import com.techyourchance.dagger2course.questions.Question
+import com.techyourchance.dagger2course.screens.common.ScreensNavigator
+import com.techyourchance.dagger2course.screens.common.dialogs.DialogsNavigator
+import com.techyourchance.dagger2course.screens.common.fragments.BaseFragment
+import kotlinx.coroutines.*
+
+class QuestionsListFragment : BaseFragment(), QuestionsListViewMvc.Listener {
+
+    private val coroutineScope = CoroutineScope(SupervisorJob() + Dispatchers.Main.immediate)
+
+    private var isDataLoaded = false
+
+    private lateinit var viewMvc: QuestionsListViewMvc
+
+    private lateinit var fetchQuestionsUseCase: FetchQuestionsUseCase
+
+    private lateinit var dialogsNavigator: DialogsNavigator
+
+    private lateinit var screensNavigator: ScreensNavigator
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+
+        fetchQuestionsUseCase = compositionRoot.fetchQuestionsUseCase
+
+        dialogsNavigator = compositionRoot.dialogsNavigator
+
+        screensNavigator = compositionRoot.screensNavigator
+    }
+
+    override fun onCreateView(inflater: LayoutInflater, container: ViewGroup?, savedInstanceState: Bundle?): View? {
+        viewMvc = QuestionsListViewMvc(LayoutInflater.from(requireContext()), container)
+        return viewMvc.rootView
+    }
+
+    override fun onStart() {
+        super.onStart()
+        viewMvc.registerListener(this)
+        if (!isDataLoaded) {
+            fetchQuestions()
+        }
+    }
+
+    override fun onStop() {
+        super.onStop()
+        coroutineScope.coroutineContext.cancelChildren()
+        viewMvc.unregisterListener(this)
+    }
+
+    override fun onRefreshClicked() {
+        fetchQuestions()
+    }
+
+    private fun fetchQuestions() {
+        coroutineScope.launch {
+            viewMvc.showProgressIndication()
+            try {
+                val result = fetchQuestionsUseCase.fetchLatestQuestions()
+                when (result) {
+                    is FetchQuestionsUseCase.Result.Success -> {
+                        viewMvc.bindQuestions(result.questions)
+                        isDataLoaded = true
+                    }
+                    is FetchQuestionsUseCase.Result.Failure -> onFetchFailed()
+                }
+            } finally {
+                viewMvc.hideProgressIndication()
+            }
+        }
+    }
+
+    private fun onFetchFailed() {
+        dialogsNavigator.showServerErrorDialog()
+    }
+
+    override fun onQuestionClicked(clickedQuestion: Question) {
+        screensNavigator.toQuestionDetails(clickedQuestion.id)
+    }
+
+
+}
